extract decoder helper methods into decoder module
